# openapi-control-vehiculos.yaml
openapi: 3.0.3
info:
  title: Control del Estado de Vehículos API
  description: |
    API REST para la gestión de turnos y evaluación técnica anual de vehículos.
    - Arquitectura: .NET 8 Web API, separación estricta FE/BE.
    - Seguridad: JWT (Bearer).
    - Catálogos reemplazan enums (EstadoTurno, EstadoVehiculo, ResultadoEvaluacion, Rol, Chequeo).
    - Reglas de negocio:
      * 8 puntos a chequear, cada uno con puntaje 1..10.
      * Total > 80 => Seguro.
      * Total < 40 => Rechequeo.
      * Cualquier punto < 5 => Rechequeo.
    Notas:
      - Los IDs son UUID (formato: uuid).
      - Los timestamps son UTC (format: date-time).
      - Se proveen códigos lógicos (codigo) en catálogos para desacoplarse de GUIDs.
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
    description: Producción
  - url: https://staging.api.example.com/v1
    description: Staging
  - url: http://localhost:5070/v1
    description: Desarrollo local
...
paths:
...
  /vehiculos/{matricula}:
    get:
      tags: [Vehículos]
      summary: Obtiene (o pre-registra) un vehículo por matrícula
      description: |
        Devuelve el vehículo si existe. Si no existe y `autoCreate=true`, lo pre-registra con datos mínimos.
      parameters:
        - $ref: '#/components/parameters/MatriculaParam'
        - in: query
          name: autoCreate
          schema: { type: boolean, default: false }
          description: Si es true, crea un registro mínimo si la matrícula no existe.
      security: []
      responses:
        '200':
          description: Vehículo encontrado o creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehiculo'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Vehículos]
      summary: Actualiza datos del vehículo (propietario, marca, modelo, año, estado)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MatriculaParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVehiculoRequest'
      responses:
        '200':
          description: Actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehiculo'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /vehiculos/{matricula}/evaluaciones:
    get:
      tags: [Vehículos, Evaluaciones]
      summary: Lista evaluaciones del vehículo por matrícula
      parameters:
        - $ref: '#/components/parameters/MatriculaParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Listado paginado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEvaluaciones'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
      description: Número de página (1-based).
    PageSizeParam:
      in: query
      name: pageSize
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
      description: Tamaño de página.
    MatriculaParam:
      in: path
      name: matricula
      required: true
      schema: { type: string, minLength: 5, maxLength: 12 }
...
  responses:
    BadRequest:
      description: Petición inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples:
            validation:
              value: { code: "validation_error", message: "Campo 'puntaje' debe estar entre 1 y 10" }
    Unauthorized:
      description: No autenticado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples: { default: { value: { code: "unauthorized", message: "Token inválido o ausente" } } }
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          examples: { default: { value: { code: "not_found", message: "No existe vehículo con la matrícula" } } }
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        traceId: { type: string, nullable: true }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              error: { type: string }
    PageMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        pageSize: { type: integer, example: 20 }
        total: { type: integer, example: 123 }
        hasNext: { type: boolean, example: true }
    PagedEvaluaciones:
      type: object
      properties:
        meta: { $ref: '#/components/schemas/PageMeta' }
        items:
          type: array
          items: { $ref: '#/components/schemas/Evaluacion' }
...
    EstadoVehiculo:
      type: object
      properties:
        id: { type: string, format: uuid }
        codigo: { type: string, example: "SEGURO" }
        nombre: { type: string, example: "Seguro" }
        orden: { type: integer, example: 2 }
    ResultadoEvaluacion:
      type: object
      properties:
        id: { type: string, format: uuid }
        codigo: { type: string, example: "RECHEQUEO" }
        nombre: { type: string, example: "Rechequeo" }
        orden: { type: integer, example: 1 }
    Chequeo:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string, example: "Frenos" }
        descripcion: { type: string, nullable: true }
        orden: { type: integer, example: 1 }
...
    Propietario:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        email: { type: string, format: email, nullable: true }
        telefono: { type: string, nullable: true }
    Vehiculo:
      type: object
      properties:
        id: { type: string, format: uuid }
        matricula: { type: string }
        marca: { type: string, nullable: true }
        modelo: { type: string, nullable: true }
        anio: { type: integer, nullable: true }
        propietarioId: { type: string, format: uuid }
        estadoVehiculoId: { type: string, format: uuid }
        propietario: { $ref: '#/components/schemas/Propietario' }
        estadoVehiculo: { $ref: '#/components/schemas/EstadoVehiculo' }
...
    EvaluacionDetalle:
      type: object
      properties:
        id: { type: string, format: uuid }
        evaluacionId: { type: string, format: uuid }
        chequeoId: { type: string, format: uuid }
        puntaje: { type: integer, minimum: 1, maximum: 10 }
        observacion: { type: string, nullable: true, maxLength: 500 }
        chequeo: { $ref: '#/components/schemas/Chequeo' }
    Evaluacion:
      type: object
      properties:
        id: { type: string, format: uuid }
        turnoId: { type: string, format: uuid }
        inspectorId: { type: string, format: uuid }
        fecha: { type: string, format: date-time }
        puntajeTotal: { type: integer, minimum: 0, maximum: 80 }
        resultadoEvaluacionId: { type: string, format: uuid }
        resultado: { $ref: '#/components/schemas/ResultadoEvaluacion' }
        detalles:
          type: array
          minItems: 8
          maxItems: 8
          items: { $ref: '#/components/schemas/EvaluacionDetalle' }

    # ---- Requests/DTO ----
...
    UpdateVehiculoRequest:
      type: object
      properties:
        marca: { type: string, nullable: true }
        modelo: { type: string, nullable: true }
        anio: { type: integer, nullable: true }
        propietarioId: { type: string, format: uuid, nullable: true }
        estadoVehiculoId: { type: string, format: uuid, nullable: true }
...
security:
  - bearerAuth: []